Assignment 2 - added extra credit part 2 (time based score)

+++ Card Match Game/EmojiCardMatchGame.swift
@@ -63,7 +63,7 @@ class EmojiCardMatchGame: ObservableObject {
         return game.cards
     }
     
-    var score: Int {
+    var score: Double {
         return game.score
     }
